refactor: Move LayoutTypeSwitch component into its own file
